Move extension field mul/sqr specialisations to their own file.
