+++ field/src/extension_field/mod.rs
@@ -6,6 +6,7 @@ pub mod algebra;
 pub mod quadratic;
 pub mod quartic;
 pub mod quintic;
+pub mod goldilocks_field;
 
 /// Optimal extension field trait.
 /// A degree `d` field extension is optimal if there exists a base field element `W`,
